requirements.txt
beautifulsoup4==4.10.0
Pillow==8.4.0
Pybooru==4.2.2
pygelbooru==0.3.1
pysaucenao==1.6.1
requests==2.25.1
syncer==1.3.0
tqdm==4.59.0
